+++ packages/pyramid_lint_test/test/assists/flutter/wrap_with_layout_builder/wrap_with_layout_builder.dart
@@ -5,6 +5,12 @@ class A extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Container();
+    return LayoutBuilder(
+      builder: (_, __) {
+        return const Center(
+          child: Placeholder(),
+        );
+      },
+    );
   }
 }