+++ packages/pyramid_lint_test/test/assists/flutter/wrap_with_layout_builder/wrap_with_layout_builder_test.dart
@@ -19,10 +19,24 @@ void main() {
       );
 
       final changes = [
-        // Container
+        // LayoutBuilder
         ...await assist.testRun(
           result,
           const SourceRange(166, 0),
+          pubspec: pubspec,
+        ),
+
+        // Center
+        ...await assist.testRun(
+          result,
+          const SourceRange(225, 0),
+          pubspec: pubspec,
+        ),
+
+        // Placeholder
+        ...await assist.testRun(
+          result,
+          const SourceRange(252, 0),
           pubspec: pubspec,
         ),
       ];